requirements.txt
altgraph==0.17.5
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
Deprecated==1.3.1
Flask==3.1.3
Flask-Limiter==4.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
macholib==1.16.4
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==26.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
python-dotenv==1.2.2
typing_extensions==4.15.0
Werkzeug==3.1.7
wrapt==2.1.2
